# refactor(modeling): linear runtime toOutlines