@z3dev 2022-02-18
someone went deep down inside geom2.toOutlines
:)