@t1u1 2022-02-18
> it's interesting that the creation of a string plus hashmap lookups can be faster then tight loops. javascript is a strange beast.

O(N) will always be faster than O(N^2) for large enough N, and for small N, the difference doesn't matter much anyway.